Refactored everything - new folders, splitted files, indexes
