+++ src/components/index.js
@@ -1,0 +1,5 @@
+// All components index
+export * from "./menu";
+export * from "./subjects";
+export * from "./timetable";
+export * from "./ui";
