+++ src/components/menu/index.js
@@ -1,0 +1,7 @@
+// Menu components
+export { default as SubjectAdder } from "./SubjectAdder";
+export { default as ServerQuery } from "./ServerQuery";
+export { default as ImportExport } from "./ImportExport";
+export { default as Settings } from "./Settings";
+export { default as Help } from "./Help";
+export { default as ImportModal } from "./ImportModal";
