+++ src/components/subjects/index.js
@@ -1,0 +1,4 @@
+// Subject components
+export { default as Subject } from "./Subject";
+export { default as SubjectModal } from "./SubjectModal";
+export { default as CourseModal } from "./CourseModal";
