+++ src/components/timetable/index.js
@@ -1,0 +1,2 @@
+// Timetable components
+export { default as Timetable } from "./Timetable";
