+++ src/components/ui/index.js
@@ -1,0 +1,3 @@
+// UI components
+export { default as Button } from "./Button";
+export { default as Footer } from "./Footer";
